requirements.txt
Flask==0.12.2
numpy==1.13.0
pandas==0.20.2
tensorflow-gpu==1.2.1
scikit-learn==0.18.2
python-dotenv==0.6.4
requests=2.13.0
